v2.0.11 增加测试数据库;更新readme

+++ movie-api/src/main/java/org/humingk/movie/api/common/vo/movie/details/CommentOfMovieDoubanVo.java
@@ -29,7 +29,7 @@ public class CommentOfMovieDoubanVo implements Serializable {
   private String urlUserDouban;
 
   /** 投票数 */
-  private Short agreeVote;
+  private Long agreeVote;
 
   /** 短评日期 */
   private String createDate;
